+++ src/test/java/dk/kb/present/TestUtil.java
@@ -119,8 +119,8 @@ public class TestUtil {
 				"kalturaID", "aVeryTrueKalturaID",
 				"startTime", "1987-05-04T14:45:00Z",
 				"endTime", "1987-05-04T16:45:00Z",
-				"ownProductionBool", "true",
-				"ownProductionCode","1000");
+				"productionCodeAllowed", "true",
+				"productionCodeValue","1000");
 		String schemaOrgJson = TestUtil.getTransformedWithAccessFieldsAdded(schemaOrgTransformer, record, injections);
 		//prettyPrintJson(schemaOrgJson);