+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.1](https://github.com/kb-dk/ds-present/releases/tag/ds-present-2.2.1) 2024-11-21
 ### Added: 
 - Added functionality to change own production threshold in configuration files, with a default value which includes own-, co- and enterprise-production for DR records.